requirements.txt
pytest~=8.3.5
spacy~=3.8.7
langdetect~=1.0.9
beautifulsoup4~=4.13.4
python-docx~=1.1.2
PyMuPDF~=1.25.5
pandas~=2.2.3